Move eventsOverlap helper to frontend
