+++ back-end/utils/misc.test.ts
@@ -3,8 +3,7 @@ import { assertEquals } from 'https://deno.land/std@0.152.0/testing/asserts.ts'
 import { stub } from "https://deno.land/std@0.220.0/testing/mock.ts"
 
 import { Tables } from '../types/db-types.ts'
-import { purchaseBreakdown, purchaseTypeAvailable, eventsOverlap } from './misc.ts'
-import dayjs from './dayjs.ts'
+import { purchaseBreakdown, purchaseTypeAvailable } from './misc.ts'
 
 Deno.test({
     name: 'purchaseBreakdown()',
@@ -192,163 +191,6 @@ Deno.test({
                 "available_to": new Date(beforeNow).valueOf()
             } as unknown as Tables['purchase_type'], { is_low_income: false }, { sales_are_open: true }),
             false
-        )
-    }
-})
-
-Deno.test({
-    name: 'eventsOverlap()',
-    fn() {
-        const baseEvent = {
-            event_id: 'test' as unknown as Tables['event']['event_id'],
-            event_site_location: 'site1' as unknown as Tables['event_site']['event_site_id'],
-            start_datetime: dayjs('2024-06-01T12:00:00Z'),
-            end_datetime: dayjs('2024-06-01T13:00:00Z')
-        }
-
-        // Same event (editing case)
-        assertEquals(
-            eventsOverlap(
-                baseEvent,
-                { ...baseEvent, start_datetime: dayjs('2024-06-01T14:00:00Z') }
-            ),
-            false
-        )
-
-        // Different locations
-        assertEquals(
-            eventsOverlap(
-                baseEvent,
-                { 
-                    ...baseEvent, 
-                    event_id: 'other' as unknown as Tables['event']['event_id'],
-                    event_site_location: 'site2' as unknown as Tables['event_site']['event_site_id']
-                }
-            ),
-            false
-        )
-
-        // No location (offsite events)
-        assertEquals(
-            eventsOverlap(
-                { ...baseEvent, event_site_location: null },
-                { 
-                    ...baseEvent, 
-                    event_id: 'other' as unknown as Tables['event']['event_id']
-                }
-            ),
-            false
-        )
-
-        // Complete overlap
-        assertEquals(
-            eventsOverlap(
-                baseEvent,
-                { 
-                    ...baseEvent, 
-                    event_id: 'other' as unknown as Tables['event']['event_id']
-                }
-            ),
-            true
-        )
-
-        // Partial overlap (end of first overlaps start of second)
-        assertEquals(
-            eventsOverlap(
-                baseEvent,
-                {
-                    ...baseEvent,
-                    event_id: 'other' as unknown as Tables['event']['event_id'],
-                    start_datetime: dayjs('2024-06-01T12:30:00Z'),
-                    end_datetime: dayjs('2024-06-01T14:00:00Z')
-                }
-            ),
-            true
-        )
-
-        // Partial overlap (start of first overlaps end of second)
-        assertEquals(
-            eventsOverlap(
-                baseEvent,
-                {
-                    ...baseEvent,
-                    event_id: 'other' as unknown as Tables['event']['event_id'],
-                    start_datetime: dayjs('2024-06-01T11:00:00Z'),
-                    end_datetime: dayjs('2024-06-01T12:30:00Z')
-                }
-            ),
-            true
-        )
-
-        // No overlap
-        assertEquals(
-            eventsOverlap(
-                baseEvent,
-                {
-                    ...baseEvent,
-                    event_id: 'other' as unknown as Tables['event']['event_id'],
-                    start_datetime: dayjs('2024-06-01T14:00:00Z'),
-                    end_datetime: dayjs('2024-06-01T15:00:00Z')
-                }
-            ),
-            false
-        )
-
-        // Buffer overlap (15 minutes)
-        assertEquals(
-            eventsOverlap(
-                baseEvent,
-                {
-                    ...baseEvent,
-                    event_id: 'other' as unknown as Tables['event']['event_id'],
-                    start_datetime: dayjs('2024-06-01T13:10:00Z'),
-                    end_datetime: dayjs('2024-06-01T14:00:00Z')
-                },
-                15
-            ),
-            true
-        )
-
-        // No end time for first event
-        assertEquals(
-            eventsOverlap(
-                { ...baseEvent, end_datetime: null },
-                {
-                    ...baseEvent,
-                    event_id: 'other' as unknown as Tables['event']['event_id'],
-                    start_datetime: dayjs('2024-06-01T13:00:00Z'),
-                    end_datetime: dayjs('2024-06-01T14:00:00Z')
-                }
-            ),
-            true
-        )
-
-        // No end time for second event
-        assertEquals(
-            eventsOverlap(
-                baseEvent,
-                {
-                    ...baseEvent,
-                    event_id: 'other' as unknown as Tables['event']['event_id'],
-                    start_datetime: dayjs('2024-06-01T11:00:00Z'),
-                    end_datetime: null
-                }
-            ),
-            true
-        )
-
-        // Both events have no end time
-        assertEquals(
-            eventsOverlap(
-                { ...baseEvent, end_datetime: null },
-                {
-                    ...baseEvent,
-                    event_id: 'other' as unknown as Tables['event']['event_id'],
-                    start_datetime: dayjs('2024-06-01T13:00:00Z'),
-                    end_datetime: null
-                }
-            ),
-            true
         )
     }
 })
