+++ front-end/src/utils.tsx
@@ -1,7 +1,9 @@
 
 import React, { ReactNode } from 'react'
 
-import { objectEntries, objectFromEntries } from '../../back-end/utils/misc'
+import { exists, objectEntries, objectFromEntries } from '../../back-end/utils/misc'
+import { Dayjs } from 'dayjs'
+import { Tables } from '../../back-end/types/db-types'
 
 export function wait(ms: number): Promise<void> {
     return new Promise(res => setTimeout(res, ms))
@@ -66,3 +68,45 @@ export function urlsToLinks(str: string): React.ReactNode[] {
 
     return segments
 }
+
+// Checks if two events overlap in location and time with buffer minutes (default 0)
+export function eventsOverlap(
+    event1: Pick<Tables['event'], 'event_id' | 'event_site_location'> & {
+      start_datetime: Dayjs,
+      end_datetime: Dayjs | null
+    },
+    event2: Pick<Tables['event'], 'event_id' | 'event_site_location'> & {
+      start_datetime: Dayjs,
+      end_datetime: Dayjs | null
+    },
+    bufferMinutes = 0
+  ): boolean {
+    if (
+      // Same event (for edit case)
+      event1.event_id === event2.event_id ||
+      // Either event has no event site location
+      !exists(event1.event_site_location) ||
+      !exists(event2.event_site_location) ||
+      // Events are at different locations
+      event1.event_site_location !== event2.event_site_location
+    ) {
+      return false
+    }
+  
+    const start1 = event1.start_datetime
+    const start2 = event2.start_datetime
+    
+    // Treat events with no end time as running indefinitely
+    if (!exists(event1.end_datetime)) {
+      return start2.unix() >= start1.subtract(bufferMinutes, 'minutes').unix()
+    }
+    if (!exists(event2.end_datetime)) {
+      return start1.unix() >= start2.subtract(bufferMinutes, 'minutes').unix()
+    }
+  
+    // Both events have end times so we check for overlap with buffer
+    return (
+      start1.unix() <= event2.end_datetime.add(bufferMinutes, 'minutes').unix() &&
+      event1.end_datetime.unix() >= start2.subtract(bufferMinutes, 'minutes').unix()
+    )
+  }
